// src/generated/api.ts
export interface components {
    schemas: {
// ...
        OverrideField: {
            /**
             * Format: int64
             * @description Allow bit flags
             */
            a: number;
            /**
             * Format: int64
             * @description Disallow bit flags
             */
            d: number;
        };
// ...
        Role: {
            /** @description Unique Id */
            _id: string;
            /** @description Role name */
            name: string;
            /** @description Permissions available to this role */
            permissions: components["schemas"]["OverrideField"];
            /**
             * @description Colour used for this role
             *
             *     This can be any valid CSS colour
             */
            colour?: string | null;
            /** @description Whether this role should be shown separately on the member sidebar */
            hoist?: boolean;
            /**
             * Format: int64
             * @description Ranking of this role
             * @default 0
             */
            rank: number;
            /** @description Role icon */
            icon?: components["schemas"]["File"] | null;
        };
// ...
        PartialRole: {
            id?: string | null;
            name?: string | null;
            permissions?: components["schemas"]["OverrideField"] | null;
            colour?: string | null;
            hoist?: boolean | null;
            /** Format: int64 */
            rank?: number | null;
            icon?: components["schemas"]["File"] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}